fix testthat errors that appear to be API-related
The problem manifested as testing failing when run
once, but not the second time. A trial-and-error
approach lead to adding a sleep() prior to the problematic
tests (test_legacy) and that seems to fix the problem. I'll
have to work the the NCI GDC to figure out what the issue is.